+++ src/app/[locale]/(route)/login/email/_hooks/useEmailLoginConfig/useEmailLoginConfig.ts
