+++ src/app/[locale]/(route)/sign-up/_hooks/useSignUpBtnClick.ts
@@ -2,7 +2,8 @@ import { useFormContext, UseFormSetFocus, useWatch } from "react-hook-form";
 import { useToast } from "@/context/ToastContext";
 import { useTranslations } from "next-intl";
 import { useEffect, useMemo, useState } from "react";
-import { useEmailErrorMessage, useEmailCheckCodeMessage } from "../_constants/SIGNUP_ERROR_MESSAGE";
+import { useEmailErrorMessage } from "./useEmailErrorMessage/useEmailErrorMessage";
+import { useEmailCheckCodeMessage } from "./useEmailCheckCodeMessage/useEmailCheckCodeMessage";
 import { throttle } from "es-toolkit/compat";
 import { useApiCheckCode, useApiSendEmail } from "@/api/fetch/auth";
 import { useErrorToast, useNicknameCheck } from "@/hooks";
