+++ src/app/[locale]/(route)/sign-up/_hooks/useSignUpInputConfig/useSignUpInputConfig.ts
